seperate interface and cleanup cmake files
